Translation suport; use tag content for fallback text (#7)
Instead of hardcoding strings that cannot be overridden by the
user, we're using Markdown-It's environment object to support
translation, either with a built-in very naive function, or
with a user-provided function. The English messages can also
be customized without touching the translation logic.

The transparent content inside <video> and <audio> tags is not
a standard accessibility feature like the ALT text; it is only
shown when the browser cannot play this content at all. The
standard recommendation is to use this text to offer a download
link. See, e.g., the examples in
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/video

This implementation does so, but it also adds the descriptive
text (if any) for good measure. For backwards compatibility, we
continue to store the descriptive text (or "Untitled ...")
in parsed.title, and it can still be accessed inside the
Handlebars template via {{title}}.